requirements.txt
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
anthropic==0.3.11
beautifulsoup4==4.12.2
lxml==4.9.3
pytz==2023.3
arxiv==1.4.8
backoff==2.2.1
jsonlines==4.0.0